remove last warnings

+++ sources/modes/TilesAttackModeManager.cpp
@@ -178,8 +178,8 @@ void TilesAttackGameModeManager::TogglePauseDisplay(bool paused) {
 }
 
 void TilesAttackGameModeManager::WillScore(int count, int type, std::vector<BranchLeaf>& out) {
-    int nb = levelToLeaveToDelete(48, limit, (type == bonus ? count * 2 : count), leavesDone);
-    for (unsigned int i=0; nb>0 && i<branchLeaves.size(); i++) {
+    int nb = levelToLeaveToDelete(48, limit, (type == (int)bonus ? count * 2 : count), leavesDone);
+    for (unsigned i=0; nb > 0 && i<branchLeaves.size(); i++) {
 		CombinationMark::markCellInCombination(branchLeaves[i].e);
         out.push_back(branchLeaves[i]);
         nb--;
